Add MagicOnion.Server.InternalTesting
