+++ tests/MagicOnion.Integration.Tests/SerializerStreamingHubTest.cs
@@ -2,7 +2,6 @@ using Grpc.Net.Client;
 using MagicOnion.Client;
 using MagicOnion.Client.DynamicClient;
 using MagicOnion.Server.Hubs;
-using MagicOnionTestServer;
 using Microsoft.AspNetCore.Mvc.Testing;
 
 namespace MagicOnion.Integration.Tests;
