+++ tests/MagicOnion.Integration.Tests/SerializerUnaryTest.cs
@@ -4,7 +4,6 @@ using MagicOnion.Client.DynamicClient;
 using MagicOnion.Serialization;
 using MagicOnion.Serialization.MessagePack;
 using MagicOnion.Server;
-using MagicOnionTestServer;
 using Microsoft.AspNetCore.Mvc.Testing;
 
 #pragma warning disable CS1998 // This async method lacks 'await' operators and will run synchronously
