+++ tests/MagicOnion.Integration.Tests/Usings.cs
@@ -4,3 +4,4 @@ global using NSubstitute;
 
 global using Grpc.Core;
 global using MessagePack;
+global using MagicOnion.Server.InternalTesting;
