+++ tests/MagicOnion.Serialization.MemoryPack.Tests/MemoryPackSerializerUnaryTest.cs
@@ -2,7 +2,6 @@ using Grpc.Net.Client;
 using MagicOnion.Client;
 using MagicOnion.Serialization.MessagePack;
 using MagicOnion.Server;
-using MagicOnionTestServer;
 using MemoryPack;
 using MessagePack;
 using Microsoft.AspNetCore.Mvc.Testing;
