+++ tests/MagicOnion.Server.Tests/_GlobalUsings.cs
@@ -1,2 +1,3 @@
 global using FluentAssertions;
+global using MagicOnion.Server.InternalTesting;
 global using Xunit;
